fix: add my config

+++ Scripts/install_pre.sh
@@ -82,18 +82,36 @@ fi
 # pacman
 
 if [ -f /etc/pacman.conf ] && [ ! -f /etc/pacman.conf.hyde.bkp ]; then
-    print_log -g "[PACMAN] " -b "modify :: " "adding extra spice to pacman..."
+    prompt_timer 120 "Would you like to add spicy to pacman? [y/n]"
+    spicy_pacman=false
 
-    # shellcheck disable=SC2154
-    [ "${flg_DryRun}" -eq 1 ] || sudo cp /etc/pacman.conf /etc/pacman.conf.hyde.bkp
-    [ "${flg_DryRun}" -eq 1 ] || sudo sed -i "/^#Color/c\Color\nILoveCandy
-    /^#VerbosePkgLists/c\VerbosePkgLists
-    /^#ParallelDownloads/c\ParallelDownloads = 5" /etc/pacman.conf
-    [ "${flg_DryRun}" -eq 1 ] || sudo sed -i '/^#\[multilib\]/,+1 s/^#//' /etc/pacman.conf
+    case "${PROMPT_INPUT}" in
+    y | Y)
+        spicy_pacman=true
+        ;;
+    n | N)
+        spicy_pacman=false
+        ;;
+    *)
+        spicy_pacman=true
+        ;;
+    esac
+    if [ "${spicy_pacman}" == true ]; then
+        print_log -g "[PACMAN] " -b "modify :: " "adding extra spice to pacman..."
 
-    print_log -g "[PACMAN] " -b "update :: " "packages..."
-    [ "${flg_DryRun}" -eq 1 ] || sudo pacman -Syyu
-    [ "${flg_DryRun}" -eq 1 ] || sudo pacman -Fy
+        # shellcheck disable=SC2154
+        [ "${flg_DryRun}" -eq 1 ] || sudo cp /etc/pacman.conf /etc/pacman.conf.hyde.bkp
+        [ "${flg_DryRun}" -eq 1 ] || sudo sed -i "/^#Color/c\Color\nILoveCandy
+        /^#VerbosePkgLists/c\VerbosePkgLists
+        /^#ParallelDownloads/c\ParallelDownloads = 5" /etc/pacman.conf
+        [ "${flg_DryRun}" -eq 1 ] || sudo sed -i '/^#\[multilib\]/,+1 s/^#//' /etc/pacman.conf
+
+        print_log -g "[PACMAN] " -b "update :: " "packages..."
+        [ "${flg_DryRun}" -eq 1 ] || sudo pacman -Syyu
+        [ "${flg_DryRun}" -eq 1 ] || sudo pacman -Fy
+    else
+        print_log -g "[PACMAN] " -stat "skipped" "pacman configuration skipped..."
+    fi
 else
     print_log -sec "PACMAN" -stat "skipped" "pacman is already configured..."
 fi
